+++ src/lib/pages/home/components/CardInfo.tsx
@@ -7,11 +7,11 @@ import type { Option } from "lib/types";
 
 const cardProps: SystemStyleObject = {
   width: "100%",
+  minH: "100%",
   padding: "16px",
   borderRadius: "8px",
   justifyContent: "space-between",
   alignItems: "center",
-  height: "100%",
   cursor: "pointer",
 };